fix: fix data module
