+++ index.html
@@ -8,6 +8,6 @@
     <title>Virtual Keyboard</title>
 </head>
 <body>
-    <script src="script.js"></script>
+    <script src="script.js" type="module"></script>
 </body>
 </html>
