+++ package.json
@@ -4,11 +4,12 @@
   "description": "",
   "main": "script.js",
   "dependencies": {
-    "eslint-config-airbnb": "^18.1.0"
+    
   },
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.1.0",
+    "eslint-config-airbnb": "^18.1.0",
     "eslint-plugin-import": "^2.20.2"
   },
   "scripts": {
