+++ script.js
@@ -1,32 +1,8 @@
-const dataObject = {
-  lowerCaseEn: ['`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', '&#8592;',
-    'Tab', 'q', 'w', 'e', 'r', 't', 'y', 'u', 'i', 'o', 'p', '[', ']', '\\',
-    'CapsLock', 'a', 's', 'd', 'f', 'g', 'h', 'j', 'k', 'l', ';', "'", 'Enter',
-    'Shift', 'z', 'x', 'c', 'v', 'b', 'n', 'm', ',', '.', '&#9650;', '/', 'Shift',
-    'Ctrl', 'Win', 'Alt', '&emsp;', 'Alt', '&#9668;', '&#9660;', '&#9658;', 'Ctrl'],
-  upperCaseEn: ['`', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', '&#8592;',
-    'Tab', 'Q', 'W', 'E', 'R', 'T', 'Y', 'U', 'I', 'O', 'P', '[', ']', '\\',
-    'CapsLock', 'A', 'S', 'D', 'F', 'G', 'H', 'J', 'K', 'L', ';', "'", 'Enter',
-    'Shift', 'Z', 'X', 'C', 'V', 'B', 'N', 'M', ',', '.', '&#9650;', '/', 'Shift',
-    'Ctrl', 'Win', 'Alt', '&emsp;', 'Alt', '&#9668;', '&#9660;', '&#9658;', 'Ctrl'],
-  lowerCaseRu: ['ё', '!', '"', '№', ';', '%', ':', '?', '*', '(', ')', '_', '+', '&#8592;',
-    'Tab', 'й', 'ц', 'у', 'к', 'е', 'н', 'г', 'ш', 'щ', 'з', 'х', 'ъ', '/',
-    'CapsLock', 'ф', 'ы', 'в', 'а', 'п', 'р', 'о', 'л', 'д', 'ж', 'э', 'Enter',
-    'Shift', 'я', 'ч', 'с', 'м', 'и', 'т', 'ь', 'б', 'ю', ',', '&#9650;', 'Shift',
-    'Ctrl', 'Win', 'Alt', '&emsp;', 'Alt', '&#9668;', '&#9660;', '&#9658;', 'Ctrl'],
-  upperCaseRu: ['Ё', '1', '2', '3', '4', '5', '6', '7', '8', '9', '0', '-', '=', '&#8592;',
-    'Tab', 'Й', 'Ц', 'У', 'К', 'Е', 'Н', 'Г', 'Ш', 'Щ', 'З', 'Х', 'Ъ', '\\',
-    'CapsLock', 'Ф', 'Ы', 'В', 'А', 'П', 'Р', 'О', 'Л', 'Д', 'Ж', 'Э', 'Enter',
-    'Shift', 'Я', 'Ч', 'С', 'М', 'И', 'Т', 'Ь', 'Б', 'Ю', '.', '&#9650;', 'Shift',
-    'Ctrl', 'Win', 'Alt', '&emsp;', 'Alt', '&#9668;', '&#9660;', '&#9658;', 'Ctrl'],
-  codes: ['Backquote', 'Digit1', 'Digit2', 'Digit3', 'Digit4', 'Digit5', 'Digit6', 'Digit7', 'Digit8',
-    'Digit9', 'Digit0', 'Minus', 'Equal', 'Backspace', 'Tab', 'KeyQ', 'KeyW', 'KeyE', 'KeyR', 'KeyT', 'KeyY',
-    'KeyU', 'KeyI', 'KeyO', 'KeyP', 'BracketLeft', 'BracketRight', 'Backslash', 'CapsLock', 'KeyA', 'KeyS',
-    'KeyD', 'KeyF', 'KeyG', 'KeyH', 'KeyJ', 'KeyK', 'KeyL', 'Semicolon', 'Quote', 'Enter', 'ShiftLeft',
-    'KeyZ', 'KeyX', 'KeyC', 'KeyV', 'KeyB', 'KeyN', 'KeyM', 'Comma', 'Period', 'ArrowUp', 'Slash',
-    'ShiftRight', 'ControlLeft', 'MetaLeft', 'AltLeft', 'Space', 'AltRight', 'ArrowLeft',
-    'ArrowDown', 'ArrowRight', 'ControlRight'],
-};
+import dataObject from './js/dataObject.js';
+
+
+const data = dataObject.dataObject;
+
 
 const isEnglish = localStorage.getItem('isEng') || 'true';
 let isCapslock = false;
@@ -35,10 +11,38 @@ if (localStorage.getItem('isEng') === null) {
   localStorage.setItem('isEng', isEnglish);
 }
 
+const capslockHandler = (elem) => {
+  const caps = elem.querySelector('.caps');
+
+  caps.addEventListener('click', () => {
+    if (!isCapslock) {
+      elem.innerHTML = '';
+      drawKeys(data.upperCaseEn, elem);
+      elem.querySelector('.caps').classList.add('active');
+      isCapslock = true;
+    } else {
+      elem.innerHTML = '';
+      drawKeys(data.lowerCaseEn, elem);
+      isCapslock = false;
+    }
+  });
+};
+const keyStyleHandler = (elem) => {
+  elem.addEventListener('mousedown', (event) => {
+    const { target } = event;
+    if (target.tagName === 'SPAN' || target.tagName === 'SECTION' || target.innerText === 'Shift' || target.innerText === 'CapsLock') return;
+    target.classList.add('active');
+  });
+  elem.addEventListener('mouseup', (event) => {
+    if (event.target.className === 'keyboard-styles' || event.target.innerText === 'Shift' || event.target.innerText === 'CapsLock') return;
+    event.target.classList.remove('active');
+  });
+};
+
 const drawKeys = (arr, elem) => {
   for (let i = 0; i < arr.length; i += 1) {
     const newKey = document.createElement('div');
-    newKey.id = dataObject.codes[i];
+    newKey.id = data.codes[i];
 
     switch (arr[i]) {
       case ('&#8592;'): newKey.classList.add('backspace');
@@ -69,14 +73,12 @@ const drawKeys = (arr, elem) => {
 
     newKey.innerHTML = arr[i];
     elem.append(newKey);
-    // newKey.append(span);
   }
 
   capslockHandler(elem);
   shiftHandler(elem);
   keyStyleHandler(elem);
 };
-
 
 const drawKeyboard = () => {
   const container = document.createElement('div');
@@ -94,17 +96,15 @@ const drawKeyboard = () => {
   keyboard.classList.add('keyboard-styles');
 
   if (localStorage.getItem('isEng') === 'true') {
-    drawKeys(dataObject.lowerCaseEn, keyboard);
+    drawKeys(data.lowerCaseEn, keyboard);
     textarea.focus();
     printOnClick(keyboard, textarea);
-    // console.log(localStorage.getItem('isEng'));
   }
 
   if (localStorage.getItem('isEng') === 'false') {
-    drawKeys(dataObject.lowerCaseRu, keyboard);
+    drawKeys(data.lowerCaseRu, keyboard);
     textarea.focus();
     printOnClick(keyboard, textarea);
-    // console.log(localStorage.getItem('isEng'));
   }
 };
 
@@ -116,30 +116,29 @@ const printOnClick = (keyboard, textfield) => {
     const position = textfield.selectionStart;
 
     switch (event.target.innerText) {
-      case ('←'):
+      case '←':
         textfield.value = textfield.value.slice(0, -1);
         break;
-      case ('Enter'): textfield.value += '\n';
+      case 'Enter': textfield.value += '\n';
         break;
-      case ('Tab'): textfield.value += '\t';
+      case 'Tab': textfield.value += '\t';
         break;
-      case ('◄'):
+      case '◄':
         textfield.focus();
         textfield.selectionStart = position - 1;
         textfield.selectionEnd = position - 1;
         break;
-      case ('►'):
+      case '►':
         textfield.focus();
         textfield.selectionStart = position + 1;
         textfield.selectionEnd = position + 1;
         break;
-      case ('CapsLock'):
-      case ('Ctrl'):
-      case ('Alt'):
-      case ('Win'):
-      case ('Shift'): break;
+      case 'CapsLock':
+      case 'Ctrl':
+      case 'Alt':
+      case 'Win':
+      case 'Shift': break;
       default: clickSelectionHandler(textfield, letter);
-      // console.log(event.target.innerText);
     }
   });
 };
@@ -155,51 +154,22 @@ const clickSelectionHandler = (textfield, str) => {
   textfield.focus();
 };
 
-const capslockHandler = (elem) => {
-  const caps = elem.querySelector('.caps');
-
-  caps.addEventListener('click', () => {
-    if (!isCapslock) {
-      elem.innerHTML = '';
-      drawKeys(dataObject.upperCaseEn, elem);
-      elem.querySelector('.caps').classList.add('active');
-      isCapslock = true;
-    } else {
-      elem.innerHTML = '';
-      drawKeys(dataObject.lowerCaseEn, elem);
-      isCapslock = false;
-    }
-  });
-};
-
 const shiftHandler = (elem) => {
   const shift = document.querySelectorAll('.shift');
 
   shift.forEach((item) => {
     item.addEventListener('mousedown', () => {
       elem.innerHTML = '';
-      if (!isCapslock) drawKeys(dataObject.upperCaseEn, elem);
-      else drawKeys(dataObject.lowerCaseEn, elem);
+      if (!isCapslock) drawKeys(data.upperCaseEn, elem);
+      else drawKeys(data.lowerCaseEn, elem);
       elem.querySelectorAll('.shift')[0].classList.add('active');
       elem.querySelectorAll('.shift')[1].classList.add('active');
     });
     item.addEventListener('mouseup', () => {
       elem.innerHTML = '';
-      if (!isCapslock) drawKeys(dataObject.lowerCaseEn, elem);
-      else drawKeys(dataObject.upperCaseEn, elem);
+      if (!isCapslock) drawKeys(data.lowerCaseEn, elem);
+      else drawKeys(data.upperCaseEn, elem);
     });
-  });
-};
-
-const keyStyleHandler = (elem) => {
-  elem.addEventListener('mousedown', (event) => {
-    const { target } = event;
-    if (target.tagName === 'SPAN' || target.tagName === 'SECTION' || target.innerText === 'Shift' || target.innerText === 'CapsLock') return;
-    target.classList.add('active');
-  });
-  elem.addEventListener('mouseup', (event) => {
-    if (event.target.className === 'keyboard-styles' || event.target.innerText === 'Shift' || event.target.innerText === 'CapsLock') return;
-    event.target.classList.remove('active');
   });
 };
 
@@ -210,12 +180,12 @@ const printOnKeypress = () => {
       event.preventDefault();
       if (isCapslock) {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.lowerCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.lowerCaseEn, document.querySelector('.keyboard-styles'));
         document.getElementById(event.code).classList.add('active');
         isCapslock = false;
       } else {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.upperCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.upperCaseEn, document.querySelector('.keyboard-styles'));
         document.querySelectorAll('DIV').forEach((item) => {
           item.classList.remove('active');
         });
@@ -224,15 +194,14 @@ const printOnKeypress = () => {
     }
 
     if (event.code === 'ShiftLeft' || event.code === 'ShiftRight') {
-      // console.log(event.code);
       event.preventDefault();
       if (document.querySelector('.caps').classList.contains('active')) {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.lowerCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.lowerCaseEn, document.querySelector('.keyboard-styles'));
         document.querySelector('.caps').classList.add('active');
       } else {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.upperCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.upperCaseEn, document.querySelector('.keyboard-styles'));
       }
     }
 
@@ -257,13 +226,13 @@ const printOnKeypress = () => {
       case '←':
         document.querySelector('TEXTAREA').value = document.querySelector('TEXTAREA').value.slice(0, -1);
         break;
-      case ('◄'):
+      case '◄':
         event.preventDefault();
         document.querySelector('TEXTAREA').focus();
         document.querySelector('TEXTAREA').selectionStart = document.querySelector('TEXTAREA').selectionStart - 1;
         document.querySelector('TEXTAREA').selectionEnd = document.querySelector('TEXTAREA').selectionEnd - 1;
         break;
-      case ('►'):
+      case '►':
         event.preventDefault();
         document.querySelector('TEXTAREA').focus();
         document.querySelector('TEXTAREA').selectionStart = document.querySelector('TEXTAREA').selectionStart + 1;
@@ -271,7 +240,6 @@ const printOnKeypress = () => {
         break;
       default:
         clickSelectionHandler(document.querySelector('TEXTAREA'), document.getElementById(event.code).innerText);
-        // document.querySelector('TEXTAREA').value += document.getElementById(event.code).innerText;
     }
 
     document.getElementById(event.code).classList.add('active');
@@ -287,11 +255,11 @@ const printOnKeypress = () => {
     if (event.code === 'ShiftLeft' || event.code === 'ShiftRight') {
       if (document.querySelector('.caps').classList.contains('active')) {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.upperCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.upperCaseEn, document.querySelector('.keyboard-styles'));
         document.querySelector('.caps').classList.add('active');
       } else {
         document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.lowerCaseEn, document.querySelector('.keyboard-styles'));
+        drawKeys(data.lowerCaseEn, document.querySelector('.keyboard-styles'));
       }
     }
     document.querySelector('TEXTAREA').focus();
@@ -300,28 +268,27 @@ const printOnKeypress = () => {
 
 const switchLanguage = () => {
   document.addEventListener('keydown', (event) => {
-    // console.log('meow');
-    if (event.ctrlKey && event.altKey) {
-      event.preventDefault();
-      if (localStorage.getItem('isEng') === 'true' && document.querySelector('.caps').classList.contains('active')) {
-        localStorage.setItem('isEng', 'false');
-        document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.upperCaseRu, document.querySelector('.keyboard-styles'));
-      } else if (localStorage.getItem('isEng') === 'false' && document.querySelector('.caps').classList.contains('active')) {
-        localStorage.setItem('isEng', 'true');
-        document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.upperCaseEn, document.querySelector('.keyboard-styles'));
-      } else if (localStorage.getItem('isEng') === 'true' && !document.querySelector('.caps').classList.contains('active')) {
-        localStorage.setItem('isEng', 'false');
-        document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.lowerCaseRu, document.querySelector('.keyboard-styles'));
-      } else if (localStorage.getItem('isEng') === 'false' && !document.querySelector('.caps').classList.contains('active')) {
-        localStorage.setItem('isEng', 'true');
-        document.querySelector('.keyboard-styles').innerHTML = '';
-        drawKeys(dataObject.lowerCaseEn, document.querySelector('.keyboard-styles'));
-      } else {
-        localStorage.setItem('isEng', 'true');
-      }
+    if (!event.ctrlKey || !event.altKey) return;
+    event.preventDefault();
+
+    if (localStorage.getItem('isEng') === 'true' && document.querySelector('.caps').classList.contains('active')) {
+      localStorage.setItem('isEng', 'false');
+      document.querySelector('.keyboard-styles').innerHTML = '';
+      drawKeys(data.upperCaseRu, document.querySelector('.keyboard-styles'));
+    } else if (localStorage.getItem('isEng') === 'false' && document.querySelector('.caps').classList.contains('active')) {
+      localStorage.setItem('isEng', 'true');
+      document.querySelector('.keyboard-styles').innerHTML = '';
+      drawKeys(data.upperCaseEn, document.querySelector('.keyboard-styles'));
+    } else if (localStorage.getItem('isEng') === 'true' && !document.querySelector('.caps').classList.contains('active')) {
+      localStorage.setItem('isEng', 'false');
+      document.querySelector('.keyboard-styles').innerHTML = '';
+      drawKeys(data.lowerCaseRu, document.querySelector('.keyboard-styles'));
+    } else if (localStorage.getItem('isEng') === 'false' && !document.querySelector('.caps').classList.contains('active')) {
+      localStorage.setItem('isEng', 'true');
+      document.querySelector('.keyboard-styles').innerHTML = '';
+      drawKeys(data.lowerCaseEn, document.querySelector('.keyboard-styles'));
+    } else {
+      localStorage.setItem('isEng', 'true');
     }
   });
 };
